+++ linux/flutter/generated_plugins.cmake
@@ -4,7 +4,9 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   open_file_linux
+  screen_retriever_linux
   url_launcher_linux
+  window_manager
   window_size
 )